Move AuthenticatorAttestationResponse and AuthenticatorAssertionResponse into entity package
