+++ powerauth-fido2-model/src/main/java/com/wultra/security/powerauth/fido2/model/request/AssertionVerificationRequest.java
@@ -20,7 +20,7 @@ package com.wultra.security.powerauth.fido2.model.request;
 
 import com.fasterxml.jackson.annotation.JsonSetter;
 import com.fasterxml.jackson.annotation.Nulls;
-import com.wultra.security.powerauth.fido2.model.response.AuthenticatorAssertionResponse;
+import com.wultra.security.powerauth.fido2.model.entity.AuthenticatorAssertionResponse;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
